Break other classes out of EmailAddress.scala
