+++ src/main/scala/uk/gov/hmrc/emailaddress/EmailAddress.scala
@@ -1,14 +1,5 @@
 package uk.gov.hmrc.emailaddress
 
-trait StringValue {
-  val value: String
-  override def toString: String = value
-}
-object StringValue {
-  import scala.language.implicitConversions
-  implicit def stringValueToString(e: StringValue): String = e.value
-}
-
 case class EmailAddress(value: String) extends StringValue {
 
   val (mailbox, domain): (EmailAddress.Mailbox, EmailAddress.Domain) = value match {
@@ -34,19 +25,5 @@ object EmailAddress {
       case EmailAddress.validDomain(_) => //
       case invalidDomain => throw new IllegalArgumentException(s"'$invalidDomain' is not a valid email domain")
     }
-  }
-}
-
-object PlayJsonFormats {
-  import play.api.libs.json._
-
-  implicit val emailAddressReads = new Reads[EmailAddress] {
-    def reads(js: JsValue): JsResult[EmailAddress] = js.validate[String].flatMap {
-      case s if EmailAddress.isValid(s) => JsSuccess(EmailAddress(s))
-      case s => JsError("not a valid email address")
-    }
-  }
-  implicit val emailAddressWrites = new Writes[EmailAddress] {
-    def writes(e: EmailAddress): JsValue = JsString(e.value)
   }
 }
